Move boudary interfaces and response objects into a seprate project register IGetAllDocumentsUseCase
