+++ HNCominoLetterPrinting.sln
@@ -19,6 +19,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UsecaseTests", "test\Usecas
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "comino-print-api", "comino-print-api\comino-print-api.csproj", "{6C9358E1-CAFB-4A92-AA3A-F5D141F09EA0}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Boundary", "src\Boundary\Boundary.csproj", "{DA2AE306-DB67-4BC2-9051-0E1405687E3D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -53,6 +55,10 @@ Global
 		{6C9358E1-CAFB-4A92-AA3A-F5D141F09EA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6C9358E1-CAFB-4A92-AA3A-F5D141F09EA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6C9358E1-CAFB-4A92-AA3A-F5D141F09EA0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DA2AE306-DB67-4BC2-9051-0E1405687E3D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DA2AE306-DB67-4BC2-9051-0E1405687E3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DA2AE306-DB67-4BC2-9051-0E1405687E3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DA2AE306-DB67-4BC2-9051-0E1405687E3D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{5BF7C470-20F6-4DC2-83FE-5C00A81CE33C} = {0BF7692B-F35E-4F5D-85BD-CFB0289BC0AF}
@@ -62,5 +68,6 @@ Global
 		{75734AB8-2B08-4DF9-874E-8527D403FCA4} = {C976BF70-517E-48B2-9A31-7FC3B9E10E9F}
 		{4FD61009-44D1-4306-8F33-4FFCF3693FC0} = {C976BF70-517E-48B2-9A31-7FC3B9E10E9F}
 		{6C9358E1-CAFB-4A92-AA3A-F5D141F09EA0} = {0BF7692B-F35E-4F5D-85BD-CFB0289BC0AF}
+		{DA2AE306-DB67-4BC2-9051-0E1405687E3D} = {0BF7692B-F35E-4F5D-85BD-CFB0289BC0AF}
 	EndGlobalSection
 EndGlobal
