+++ comino-print-api/comino-print-api.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
         <TargetFramework>netcoreapp2.1</TargetFramework>
@@ -17,8 +17,8 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\src\Boundary\Boundary.csproj" />
       <ProjectReference Include="..\src\Lambda\Lambda.csproj" />
-      <ProjectReference Include="..\src\Usecases\Usecases.csproj" />
     </ItemGroup>
