+++ src/Boundary/Responses/GetAllDocumentsResponse.cs
