+++ src/Boundary/UseCaseInterfaces/IGetAllDocuments.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using comino_print_api.Models;
 
-namespace comino_print_api.UseCaseInterFaces
+namespace Boundary.UseCaseInterfaces
 {
     public interface IGetAllDocuments
     {
