+++ src/Usecases/Interfaces/IGetDetailsOfDocumentForProcessing.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using Usecases.Domain;
 
-namespace Usecases.UseCaseInterfaces
+namespace Usecases.Interfaces
 {
     public interface IGetDetailsOfDocumentForProcessing
     {
