+++ src/Usecases/Interfaces/ISaveRecordsToLocalDatabase.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Usecases.Domain;
 
-namespace Usecases.UseCaseInterfaces
+namespace Usecases.Interfaces
 {
     public interface ISaveRecordsToLocalDatabase
     {
