+++ src/Usecases/UntestedParsers/IronPDFParser.cs
@@ -1,6 +1,7 @@
 using System.Text;
+using Boundary.UseCaseInterfaces;
 using IronPdf;
-using Usecases.UseCaseInterfaces;
+using Usecases.Interfaces;
 
 namespace UseCases.UntestedParsers
 {
