--- src/Usecases/UseCaseInterfaces/ISaveToLocalDatabase.cs
@@ -1,7 +1,0 @@
-namespace AwsDotnetCsharp.UsecaseInterfaces
-{
-    public interface ISaveToLocalDatabase
-    {
-        
-    }
-}
