+++ src/Usecases/Usecases.csproj
@@ -13,4 +13,8 @@
       <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\Boundary\Boundary.csproj" />
+    </ItemGroup>
+
 </Project>
